MembershipsService, RolesService, TeamsService, and UsersServices were changed to MembershipService, RoleService, TeamService, and UserServices, in order to follow the java standard of classes' name where the classes or interfaces have to be in singular
TeamsClient, and UsersClient where changed to TeamRestClient, and UserRestClient in order
to follow the java standard of classes' name where the classes or interfaces have to be in singular and marks in a explicity way that those clients are rest clients because those
are using restTemplate object.

MembershipsServiceImpl, RolesServiceImpl, TeamsServiceImpl, and UsersServiceImpl were changed to MembershipServiceImpl, RoleServiceImpl, TeamServiceImpl, and UserServiceImpl, in order
to follow the java standard of classes' name where the classes or interfaces have to be in singular

MembershipsApi, RolesApi, TeamsApi, and UsersApi were changed to MembershipApi, RoleApi, TeamApi, and UserApi , in order
to follow the java standard of classes' name where the classes or interfaces have to be in singular

MembershipsRestController, RolesRestController, TeamsRestController, and UsersRestController were changed to MembershipsRestController, RoleRestController, TeamRestController, and UserRestController , in order
to follow the java standard of classes' name where the classes or interfaces have to be in singular

MembershipsApiTests, RolesApiTest, RolesServiceTest, and UsersServiceTest were changed to MembershipApiTests, RoleApiTest, RoleServiceTest, and UserServiceTest, in order
to follow the java standard of classes' name where the classes or interfaces have to be in singular

In MembershipServiceImpl, and RoleServiceImpl, I removed the unused @Log4j2 annotation, it will create an static final object automatically that is not used
inside the service,  (I removed the unused imports after remove @Log4j2)

In MembershipServiceImpl, RoleServiceImpl, and TeamServiceImpl, UserServices I added the @RequiredArgsConstructor annotation,
this is for reduce the lines of code and removed boilerplate code (I removed the unused imports after remove @Autowired)

In MembershipServiceImpl.assignRoleToMembership, I become the local variable roleId final,
because that variable will not be changed during the method, this improve
the readability of the code and help to the compiler in the moment
to compile de code

In MembershipServiceImpl.assignRoleToMembership, I changed the parameter name of m to membership
it will help the readability of the code

In MembershipServiceImpl.getMemberships,I changed the parameter name of rid to roleId
it will help the readability of the code

In RolesServiceImpl, I removed the DEFAULT_ROLE attibute because it is not used, and it was removed the private method getDefaultRole, because
it is not used

In RoleServiceImpl.CreateRole, I changed by RoleServiceImpl.createRole in order to follow the java standard of methods' name
In RoleServiceImpl.GetRole, I changed by RoleServiceImpl.getRole in order to follow the java standard of methods' name
In RoleServiceImpl.GetRoles, I changed by RoleServiceImpl.getRoles in order to follow the java standard of methods' name
all of them were changed in the interface (RoleService) as well.
all of them were changed in the controller class as well.
all of them were changed in the Test class as well.

In RoleServiceImpl.CreateRole, I changed the parameter name of r to role
it will help the readability of the code

In RoleServiceImpl.getRole, I changed the parameter name of rid to roleId
it will help the readability of the code.

In TeamServiceImpl, I become the atribute from teamsClient to teamRestClient,
in order to follow that the atribbute has to be named in singular if it is not a collection and
to indicate that it is a rest client

In UserServiceImpl, I become the atribute from usersClient to userRestClient,
in order to follow that the atribbute has to be named in singular if it is not a collection and
to indicate that it is a rest client

In TeamServiceImpl.getTeam, I changed the parameter name of id to teamId
it will help the readability of the code.

In UserServiceImpl.getUser, I changed the parameter name of id to userId
it will help the readability of the code.

In RoleRestController.getRoles, and RoleRestController.getRole were changed from postmapping annotation to getmapping annotation, this
because following the restfull philosophy, for actions where just is necessary to fetch data, the service should use GET http verb.

In MembershipRestController.getMemberships was changed from postmapping annotation to getmapping annotation, this
because following the restfull philosophy, for actions where just is necessary to fetch data, the service should use GET http verb.
getMemberships

In RoleRestController.getMember, the status code was changed to 201, it means that the record was created.
In MembershipRestController.MembershipRestController, the status code was changed to 201, it means that the record was created.

In MembershipServiceImpl.assignRoleToMembership it was changed the logic to validate if the membership exist for the role, team and user
In MembershipRepository  it was created a new method MembershipRepository.findByUserIdAndTeamIdAndRoleId to validate if exist a membership with one specific
userId, TeamId and RoleId

In MembershipRepository.findByUserIdAndTeamId it was changed to return a list, because taking into account the logic of the database, it can return
several records

In MembershipServiceTest.shouldCreateMembership and MembershipServiceTest.shouldFailToCreateMembershipWhenItExists were changed the method findByUserIdAndTeamId for
MembershipRepository.findByUserIdAndTeamIdAndRoleId and also it was necessary to mock the team.getTeam, because the logic of the service was changed.

It was created a new Exception  named InvalidLogicException, it will represent one exeception with the status code 400, it was added in MembershipServiceImpl.assignRoleToMembership
to validate if one userId is part of a team, also it was added the logic to validate if a user is part of a team.

about the new endpoint to create /v1/roles/search (search roles by teamId and userId as request params), I created the new endpoint in the RoleController class, and I added
the new logic inside RoleServiceImpl, I created the new method in the service and implement it in the class, I follow the same standard of validations, and I update
the necessary unit test.